+++ src/main/java/ru/yandex/practicum/main/Main.java
@@ -1,11 +1,11 @@
-package ru.yandex.practicum.main;
+package main.java.ru.yandex.practicum.main;
 
-import ru.yandex.practicum.main.model.Epic;
-import ru.yandex.practicum.main.model.Status;
-import ru.yandex.practicum.main.model.SubTask;
-import ru.yandex.practicum.main.model.Task;
-import ru.yandex.practicum.main.service.Manager;
-import ru.yandex.practicum.main.service.TaskManager;
+import main.java.ru.yandex.practicum.main.model.Epic;
+import main.java.ru.yandex.practicum.main.model.Status;
+import main.java.ru.yandex.practicum.main.model.SubTask;
+import main.java.ru.yandex.practicum.main.model.Task;
+import main.java.ru.yandex.practicum.main.service.Manager;
+import main.java.ru.yandex.practicum.main.service.TaskManager;
 
 import java.util.List;
