+++ src/main/java/ru/yandex/practicum/main/model/Status.java
@@ -1,0 +1,7 @@
+package main.java.ru.yandex.practicum.main.model;
+
+public enum Status {
+    NEW,
+    IN_PROGRESS,
+    DONE
+}
