--- src/ru/yandex/practicum/main/model/Status.java
@@ -1,7 +1,0 @@
-package ru.yandex.practicum.main.model;
-
-public enum Status {
-    NEW,
-    IN_PROGRESS,
-    DONE
-}
